Some changes addressing PR comments, still likely some additional changes to come
